requirements.txt
django==2
pillow
django-filter
django-lazysignup